+++ lib/image/classification.ex
@@ -71,6 +71,8 @@ if Image.bumblebee_configured?() do
     @default_classifier [
       model: {:hf, "microsoft/resnet-50"},
       featurizer: {:hf, "microsoft/resnet-50"},
+      model_options: [],
+      featurizer_options: [],
       name: Image.Classification.Server,
       autostart: true
     ]
@@ -83,8 +85,8 @@ if Image.bumblebee_configured?() do
 
     ### Arguments
 
-    * `configuration` is a keyword list.The default is
-      `Application.get_env(:image, :classifier, [])`.
+    * `configuration` is a keyword list of options which
+      are merged over the default configuration.
 
     ### Configuration keys
 
@@ -94,12 +96,11 @@ if Image.bumblebee_configured?() do
     * `:featurizer` is any supported machine learning model for image
       featurization supported by Bumblebee.
 
-    * `:featurizer_module` is the module name of a featurizer known
-      to Bumblee. Bumblebee will attempt to infer the featurizer module
-      name but it may not always be able to do so. Specify only if
-      Bumblebee is unable to resolve a featurizer module name.
-      [Bumblebee.Vision.VitFeaturizer](https://hexdocs.pm/bumblebee/Bumblebee.Vision.VitFeaturizer.html)
-      may be an appropriate choice when a featurizer module is required.
+    * `:model_options` is a keyword list of options that
+      are passed to `Bumblebee.load_model/2`.
+
+    * `:featurizer_options` is a keyword list of options that
+      are passed to `Bumblebee.load_featurizer/2`.
 
     * `:name` is the name given to the classification process when
       it is started.
@@ -111,6 +112,8 @@ if Image.bumblebee_configured?() do
     [
       model: {:hf, "microsoft/resnet-50"},
       featurizer: {:hf, "microsoft/resnet-50"},
+      model_options: [],
+      featurizer_options: [],
       name: Image.Classification.Server
     ]
     ```
@@ -121,11 +124,13 @@ if Image.bumblebee_configured?() do
       Application.ensure_all_started(:exla)
       classifier = Keyword.merge(@default_classifier, classifier)
 
-      model = classifier[:model]
-      featurizer = classifier[:featurizer]
-      featurizer_module = classifier[:featurizer_module]
+      model = Keyword.fetch!(classifier, :model)
+      model_options = Keyword.fetch!(classifier, :model_options)
 
-      case Image.Classification.serving(model, featurizer, featurizer_module) do
+      featurizer = Keyword.fetch!(classifier, :featurizer)
+      featurizer_options = Keyword.fetch!(classifier, :featurizer_options)
+
+      case Image.Classification.serving(model, model_options, featurizer, featurizer_options) do
         {:error, error} ->
           {:error, error}
 
@@ -135,22 +140,14 @@ if Image.bumblebee_configured?() do
     end
 
     @doc false
-    def serving(model, featurizer, featurizer_module) do
-      with {:ok, model_info} <- Bumblebee.load_model(model),
-           {:ok, featurizer} = load_featurizer(featurizer, featurizer_module) do
+    def serving(model, model_options, featurizer, featurizer_options) do
+      with {:ok, model_info} <- Bumblebee.load_model(model, model_options),
+           {:ok, featurizer} = Bumblebee.load_featurizer(featurizer, featurizer_options) do
         Bumblebee.Vision.image_classification(model_info, featurizer,
           compile: [batch_size: 10],
           defn_options: [compiler: EXLA]
         )
       end
-    end
-
-    defp load_featurizer(featurizer, nil = _featurizer_module) do
-       Bumblebee.load_featurizer(featurizer)
-    end
-
-    defp load_featurizer(featurizer, featurizer_module) do
-       Bumblebee.load_featurizer(featurizer, module: featurizer_module)
     end
 
     @doc """